Drop inih subproject, always use system inih
rwmem depended on inih that was customized (no multiline support). But
if using system's inih, the multiline support was compiled in.
Furthermore buildroot doesn't like subprojects, and using inih is an
extra feature for rwmem.

So: Drop inih subproject, and always use system inih. Also make inih
optional, as it's not really needed.

This means that the ini files may need changes, as the text has to start
from the beginning of the line.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>